[DP-35] - перенес ProgressButton в CustomButtons
