+++ DavaiPosmotrim/Application/AppCoordinator.swift
@@ -14,7 +14,7 @@ final class AppCoordinator: BaseCoordinator {
         } else if UserDefaults.standard.value(forKey: Resources.Authentication.savedNameUserDefaultsKey) == nil {
             showAuthFlow()
         } else {
-            showMainFlow()
+            showSelectionMoviesFlow()
         }
     }
