+++ DavaiPosmotrim/UI/CustomButtons.swift
@@ -13,6 +13,14 @@ final class CustomButtons: UIView {
 
     private let tappedButtonAlpha = 0.7
     private let unTappedButtonAlpha = 1.0
+    private let progressLayer = CAShapeLayer()
+    private var progress: CGFloat = 0 {
+        didSet {
+            progressLayer.strokeEnd = progress
+        }
+    }
+
+    var onProgressComplete: (() -> Void)?
 
     // MARK: - Lazy properties
 
@@ -43,6 +51,16 @@ final class CustomButtons: UIView {
         return button
     }()
 
+    lazy var progressButton: UIButton = {
+        let button = UIButton()
+        button.backgroundColor = .baseTertiaryAccent
+        button.setTitleColor(.whiteText, for: .normal)
+        button.titleLabel?.font = .textButtonMediumFont
+        button.layer.cornerRadius = 12
+        button.layer.masksToBounds = true
+        return button
+    }()
+
     // MARK: - Initializers
 
     override init(frame: CGRect) {
@@ -52,6 +70,11 @@ final class CustomButtons: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+
+    override func layoutSubviews() {
+            super.layoutSubviews()
+            setupProgressLayer()
+        }
 
     // MARK: - Public methods
 
@@ -74,6 +97,66 @@ final class CustomButtons: UIView {
         button.addTarget(self, action: #selector(unTap), for: .touchUpInside)
     }
 
+    private func setupProgressLayer() {
+        progressLayer.removeFromSuperlayer()
+        let width = bounds.size.width
+        let height = bounds.size.height
+        let cornerRadius: CGFloat = 12
+        let startPoint = CGPoint(x: width / 2, y: 0)
+        let progressPath = UIBezierPath()
+        progressPath.move(to: startPoint)
+        progressPath.addLine(to: CGPoint(x: width - cornerRadius, y: 0))
+        progressPath.addArc(
+            withCenter: CGPoint(x: width - cornerRadius, y: cornerRadius),
+            radius: cornerRadius,
+            startAngle: -CGFloat.pi / 2,
+            endAngle: 0,
+            clockwise: true
+        )
+        progressPath.addLine(to: CGPoint(x: width, y: height - cornerRadius))
+        progressPath.addArc(
+            withCenter: CGPoint(x: width - cornerRadius, y: height - cornerRadius),
+            radius: cornerRadius,
+            startAngle: 0,
+            endAngle: CGFloat.pi / 2,
+            clockwise: true
+        )
+        progressPath.addLine(to: CGPoint(x: cornerRadius, y: height))
+        progressPath.addArc(
+            withCenter: CGPoint(x: cornerRadius, y: height - cornerRadius),
+            radius: cornerRadius,
+            startAngle: CGFloat.pi / 2,
+            endAngle: CGFloat.pi,
+            clockwise: true
+        )
+        progressPath.addLine(to: CGPoint(x: 0, y: cornerRadius))
+        progressPath.addArc(
+            withCenter: CGPoint(x: cornerRadius, y: cornerRadius),
+            radius: cornerRadius,
+            startAngle: CGFloat.pi,
+            endAngle: -CGFloat.pi / 2,
+            clockwise: true
+        )
+        progressPath.close()
+        progressLayer.path = progressPath.cgPath
+        progressLayer.strokeColor = UIColor.baseSecondaryAccent.cgColor
+        progressLayer.lineWidth = 3
+        progressLayer.fillColor = UIColor.clear.cgColor
+        progressLayer.lineCap = .round
+        progressLayer.strokeEnd = 0
+        layer.addSublayer(progressLayer)
+    }
+
+    func startProgress(duration: TimeInterval) {
+        let animation = CABasicAnimation(keyPath: "strokeEnd")
+        animation.toValue = 1
+        animation.duration = duration
+        animation.fillMode = .forwards
+        animation.isRemovedOnCompletion = false
+        animation.delegate = self
+        progressLayer.add(animation, forKey: "progressAnim")
+    }
+
     // MARK: - Handlers
 
     @objc private func tap() {
@@ -82,5 +165,15 @@ final class CustomButtons: UIView {
 
     @objc private func unTap() {
         alpha = unTappedButtonAlpha
+    }
+}
+
+// MARK: - CAAnimationDelegate
+
+extension CustomButtons: CAAnimationDelegate {
+    func animationDidStop(_ anim: CAAnimation, finished flag: Bool) {
+        if flag {
+            onProgressComplete?()
+        }
     }
 }
